image_remapping/requirements.txt
# Core Dependencies
numpy>=1.21.0
opencv-python>=4.5.0
scipy>=1.7.0
pillow>=8.0.0

# Web Interface
gradio>=4.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Data Processing
pandas>=1.3.0

# Optional Dependencies for Enhanced Functionality
# Uncomment as needed:

# scikit-image>=0.18.0          # Additional image processing
# scikit-learn>=1.0.0           # Machine learning utilities
# opencv-contrib-python>=4.5.0  # Extended OpenCV modules
# plotly>=5.0.0                 # Interactive plotting
# dash>=2.0.0                   # Alternative web framework

# Development Dependencies (for contributors)
# pytest>=6.0.0                 # Testing framework
# pytest-cov>=2.12.0            # Coverage reporting
# black>=21.0.0                 # Code formatting
# flake8>=3.9.0                 # Linting
# mypy>=0.910                   # Type checking

# Documentation Dependencies
# sphinx>=4.0.0                 # Documentation generation
# sphinx-rtd-theme>=0.5.0       # Documentation theme

---

# Image Remapping Suite - Comprehensive Toolkit

## System Requirements

- Python 3.8 or higher
- At least 4GB RAM (8GB recommended for large images)
- Modern web browser for Gradio interface

## Installation Instructions

### Standard Installation
```bash
pip install -r requirements.txt
```

### Development Installation
```bash
# Install with development dependencies
pip install -r requirements.txt
pip install pytest pytest-cov black flake8 mypy

# Install in development mode
pip install -e .
```

### Docker Installation (Optional)
```dockerfile
FROM python:3.9-slim

WORKDIR /app
COPY requirements.txt .
RUN pip install -r requirements.txt

COPY . .
EXPOSE 7860

CMD ["python", "main.py"]
```

## Platform-Specific Notes

### Windows
- Ensure Microsoft Visual C++ redistributables are installed
- Consider using Anaconda for easier OpenCV installation

### macOS  
- May require Xcode command line tools for some dependencies
- Use Homebrew for system-level dependencies if needed

### Linux
- Most distributions include required system libraries
- Ubuntu/Debian: `sudo apt-get install python3-dev`
- CentOS/RHEL: `sudo yum install python3-devel`

## Troubleshooting

### Common Issues

1. **OpenCV Installation Problems**
   ```bash
   # Try alternative installation
   pip uninstall opencv-python
   pip install opencv-python-headless
   ```

2. **Gradio Port Conflicts**
   ```bash
   # Specify custom port
   python main.py --port 7861
   ```

3. **Memory Issues with Large Images**
   - Reduce image dimensions in config/settings.py
   - Increase system virtual memory
   - Use lower grid densities for initial testing

4. **Import Errors**
   ```bash
   # Ensure proper Python path
   export PYTHONPATH="${PYTHONPATH}:/path/to/image_remapping_suite"
   ```

### Performance Optimization

For better performance, consider installing:
```bash
# Intel Math Kernel Library acceleration
pip install mkl

# CUDA acceleration (if NVIDIA GPU available)  
pip install opencv-python-cuda

# Numba for JIT compilation
pip install numba
```